#ue5 - Revert function param change in BP GetSelectionBounds() because C++ uses will complain. Need to deprecate and replace with another method for BPs.
#jira UE-160830
#rb none
#preflight 630e8c1df92416fb92a70f11

[CL 21730584 by Zak Middleton in ue5-main branch]

+++ Engine/Source/Editor/Blutility/Private/EditorUtilityLibrary.cpp
@@ -210,7 +210,7 @@ TArray<AActor*> UEditorUtilityLibrary::GetSelectionSet()
 	return Result;
 }
 
-void UEditorUtilityLibrary::GetSelectionBounds(FVector& Origin, FVector& BoxExtent, double& SphereRadius)
+void UEditorUtilityLibrary::GetSelectionBounds(FVector& Origin, FVector& BoxExtent, float& SphereRadius)
 {
 	bool bFirstItem = true;
 
@@ -234,7 +234,7 @@ void UEditorUtilityLibrary::GetSelectionBounds(FVector& Origin, FVector& BoxExte
 
 	Origin = Extents.Origin;
 	BoxExtent = Extents.BoxExtent;
-	SphereRadius = Extents.SphereRadius;
+	SphereRadius = (float)Extents.SphereRadius; // TODO: LWC: should be double, but need to deprecate function and replace for old C++ references to continue working.
 }
 
 TArray<UObject*> UEditorUtilityLibrary::GetSelectedAssets()
